+++ components/expansion-panel/src/vwc-expansion-panel.ts
@@ -1,5 +1,6 @@
 import '@vonage/vvd-core';
 import '@vonage/vwc-icon';
+import '@material/mwc-ripple/mwc-ripple.js';
 import {
 	customElement,
 	eventOptions,
@@ -8,11 +9,11 @@ import {
 	queryAsync,
 	TemplateResult
 } from 'lit-element';
-// import '@material/mwc-ripple/mwc-ripple';
 import type { Ripple } from '@material/mwc-ripple';
 import { RippleHandlers } from '@material/mwc-ripple/ripple-handlers.js';
 import { VWCExpansionPanelBase } from './vwc-expansion-panel-base.js';
 import { style } from './vwc-expansion-panel.css.js';
+import {nothing} from 'lit-html';
 
 declare global {
 	interface HTMLElementTagNameMap {
@@ -23,59 +24,75 @@ declare global {
 const iconSets = ['chevron', 'binary'];
 export type IndicatorIconSets = typeof iconSets;
 
+const VALID_HEADER_VALUES = [2, 3, 4, 5, 6];
+
+function isValidHeaderValue(headerValue: string | number) {
+	return VALID_HEADER_VALUES.includes(Number(headerValue));
+}
+
 @customElement('vwc-expansion-panel')
 export class VWCExpansionPanel extends VWCExpansionPanelBase {
+	safeHtml = html;
 	static override styles = style;
 
 	/**
 	 * @deprecated use "heading" instead
 	 */
-	@property({ type: String, reflect: true })
+	@property({type: String, reflect: true})
 		header = '';
 
 	/**
 	 * The heading of the expanded panel
 	 */
-	@property({ type: String, reflect: true })
+	@property({type: String, reflect: true})
 		heading = '';
 
-	@property({ type: String, reflect: true })
+	@property({type: String, reflect: true})
 		icon = '';
 
-	@property({ type: String, reflect: true })
+	@property({type: String, reflect: true})
 		indicatorIconSet: IndicatorIconSets[number] = 'chevron';
+
+	@property({ type: String, reflect: true })
+		meta = '';
 
 	@property({ type: Boolean, reflect: true })
 		dense = false;
 
-	@property({ type: Boolean, reflect: true })
+	@property({type: Boolean, reflect: true})
 		leadingToggle = false;
 
-	@property({ type: Boolean, reflect: true })
+	@property({type: Boolean, reflect: true})
 		noRipple = false;
-
 	@queryAsync('mwc-ripple') ripple!: Promise<Ripple>;
-
 	protected rippleHandlers = new RippleHandlers(() => {
 		return this.ripple;
 	});
-
-	protected toggleOpen(): void {
-		this.open = !this.open;
-	}
+	@property({type: String, reflect: true, attribute: 'heading-level'})
+	private headingLevel = '3';
 
 	override openChanged(isOpen: boolean): void {
 		super.openChanged(isOpen);
 		this.toggleAttribute('open', isOpen);
 	}
 
-	protected renderRipple(): TemplateResult | string {
-		return !this.noRipple ? html`<mwc-ripple></mwc-ripple>` : '';
+	protected toggleOpen(): void {
+		this.open = !this.open;
 	}
 
-	protected override render(): TemplateResult {
+	protected renderRipple(): TemplateResult | string {
+		return !this.noRipple ? html`
+			<mwc-ripple></mwc-ripple>` : '';
+	}
+
+	private renderMeta(): TemplateResult | unknown {
+		return this.meta ? html`<span class="meta">${this.meta}</span>` : nothing;
+	}
+
+	protected renderHeaderButton(): TemplateResult {
+
 		return html`
-			<button class="expansion-panel-header"
+			<button class="expansion-panel-button" id="expansion-panel"
 				@mousedown="${this.handleRippleActivate}"
 				@mouseenter="${this.handleRippleMouseEnter}"
 				@mouseleave="${this.handleRippleMouseLeave}"
@@ -83,11 +100,11 @@ export class VWCExpansionPanel extends VWCExpansionPanelBase {
 		this.toggleOpen();
 		this.handleRippleActivate;
 	}}"
-				@touchend="${this.handleRippleDeactivate}"
-				@touchcancel="${this.handleRippleDeactivate}"
-				@click=${() => this.toggleOpen()}
-				?aria-expanded=${this.open}
-				aria-controls="content"
+							@touchend="${this.handleRippleDeactivate}"
+							@touchcancel="${this.handleRippleDeactivate}"
+							@click=${() => this.toggleOpen()}
+							?aria-expanded=${this.open}
+							aria-controls="content"
 			>
 				${this.renderRipple()}
 				<span class="leading-icon">
@@ -95,23 +112,37 @@ export class VWCExpansionPanel extends VWCExpansionPanelBase {
 						${this.renderIconOrToggle()}
 					</slot>
 				</span>
-				${this.heading || this.header}
+				<span class="heading-text">${this.heading || this.header}</span>
+				${this.renderMeta()}
 				<span class="trailing-icon">
 					<slot name="trailingIcon">
 						${!this.leadingToggle ? this.renderToggle() : ''}
 					</slot>
 				</span>
 			</button>
+		`;
+	}
+
+	protected renderPanelHeader(): TemplateResult | string {
+		if (!isValidHeaderValue(this.headingLevel)) this.headingLevel = '3';
+		return eval(`this.safeHtml\`<h${this.headingLevel} class="expansion-panel-header">\${this.renderHeaderButton()}</h${this.headingLevel}>\``);
+	}
+
+	protected override render(): TemplateResult {
+		return html`
+			${this.renderPanelHeader()}
 			<div id="content" class="expansion-panel-body">
 				<slot></slot>
 			</div>`;
+
 	}
 
 	protected renderIconOrToggle(): TemplateResult | string {
 		if (this.leadingToggle) {
 			return this.renderToggle();
 		} else if (this.icon) {
-			return html`<vwc-icon type="${this.icon}" size="medium"></vwc-icon>`;
+			return html`
+				<vwc-icon type="${this.icon}" size="medium"></vwc-icon>`;
 		} else {
 			return '';
 		}
@@ -132,7 +163,7 @@ export class VWCExpansionPanel extends VWCExpansionPanelBase {
 		`;
 	}
 
-	@eventOptions({ passive: true })
+	@eventOptions({passive: true})
 	private handleRippleActivate(evt?: Event) {
 		const onUp = () => {
 			window.removeEventListener('mouseup', onUp);