+++ ui-tests/index.ts
@@ -142,10 +142,13 @@ async function setDevServer() {
 		...webpackConfig,
 		mode: 'development'
 	});
+
 	const devServerOptions = {
 		...webpackConfig.devServer,
 		open: true
 	};
+
+	//@ts-ignore
 	const devServer = new WebpackDevServer(devServerOptions, compiler);
 
 	devServer.listen(webpackConfig.devServer.port, '127.0.0.1', () => {