+++ components/elevation/CHANGELOG.md
@@ -1,3 +1,67 @@
+## [2.25.10](https://github.com/vonage/vivid/compare/v2.25.9...v2.25.10) (2022-03-07)
+
+
+
+## [2.25.9](https://github.com/vonage/vivid/compare/v2.25.8...v2.25.9) (2022-02-06)
+
+
+### Bug Fixes
+
+* **workspace:** resolve missing dependencies ([#1230](https://github.com/vonage/vivid/issues/1230)) ([b52e418](https://github.com/vonage/vivid/commit/b52e4180e7bcfbbb338db2c96b29f12d60ca241b))
+
+
+
+## [2.25.8](https://github.com/vonage/vivid/compare/v2.25.7...v2.25.8) (2022-02-05)
+
+
+
+## [2.25.7](https://github.com/vonage/vivid/compare/v2.25.6...v2.25.7) (2022-02-02)
+
+
+
+## [2.25.6](https://github.com/vonage/vivid/compare/v2.25.5...v2.25.6) (2022-01-30)
+
+
+
+## [2.25.5](https://github.com/vonage/vivid/compare/v2.25.3...v2.25.5) (2022-01-25)
+
+
+
+## [2.25.3](https://github.com/vonage/vivid/compare/v2.25.2...v2.25.3) (2022-01-24)
+
+
+
+## [2.25.2](https://github.com/vonage/vivid/compare/v2.25.1...v2.25.2) (2022-01-24)
+
+
+
+## [2.25.1](https://github.com/vonage/vivid/compare/v2.25.0...v2.25.1) (2022-01-23)
+
+
+### Bug Fixes
+
+* **elevation:** sharp or rounded shape ([#1185](https://github.com/vonage/vivid/issues/1185)) ([545c6b2](https://github.com/vonage/vivid/commit/545c6b2a7ab1e38c29cba301ae063b6d9f2767b0))
+
+
+
+# [2.25.0](https://github.com/vonage/vivid/compare/v2.24.1...v2.25.0) (2022-01-11)
+
+
+### Bug Fixes
+
+* **elevation:** new elevation dp ([#1177](https://github.com/vonage/vivid/issues/1177)) ([ddc4c52](https://github.com/vonage/vivid/commit/ddc4c5289111b8436388a0221142e925f6a05b80))
+* **elevation:** remove containment ([#1188](https://github.com/vonage/vivid/issues/1188)) ([57e40b3](https://github.com/vonage/vivid/commit/57e40b364b76761311a131b3ff4d0ed034d2df7d))
+
+
+
+## [2.24.1](https://github.com/vonage/vivid/compare/v2.24.0...v2.24.1) (2021-12-23)
+
+
+
+# [2.24.0](https://github.com/vonage/vivid/compare/v2.23.0...v2.24.0) (2021-12-22)
+
+
+
 # [2.23.0](https://github.com/vonage/vivid/compare/v2.22.1...v2.23.0) (2021-12-14)